hw2/requirements.txt
Pillow
TensorFlow
Keras
